[chore] 배치 모듈 배포

+++ photi-apis/enduser/src/main/kotlin/com/photi/apis/enduser/controller/user/dto/response/FindChallengeHistoryResponse.kt
@@ -31,7 +31,7 @@ data class FindChallengeHistoryResponse(
             challengeHistory.username,
             challengeHistory.imageUrl,
             challengeHistory.feedCount,
-            challengeHistory.endedChallengeCount,
+            challengeHistory.endedChallengeCount.toInt(),
             challengeHistory.signInDate.toLocalDate(),
         )
     }
